+++ src/main/java/com/system/fcfs/event/config/RedisConfig.java
@@ -1,6 +1,5 @@
-package com.system.fcfs.event.redis.config;
+package com.system.fcfs.event.config;
 
-import com.system.fcfs.event.redis.properties.RedisProperty;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
