--- src/main/java/com/system/fcfs/event/dto/GetWinnerRequestDTO.java
@@ -1,4 +1,0 @@
-package com.system.fcfs.event.dto;
-
-public class GetWinnerRequestDTO {
-}
