--- src/main/java/com/system/fcfs/event/dto/PostEventRequestDTO.java
@@ -1,17 +1,0 @@
-package com.system.fcfs.event.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class PostEventRequestDTO {
-    private String userId;
-    private Long rank;
-    private String timestamp;
-    private String eventName;
-
-    public void setRank(Long rank) {
-        this.rank = rank;
-    }
-}
