+++ src/main/java/com/system/fcfs/event/dto/request/GetWinnerRequestDTO.java
@@ -1,0 +1,4 @@
+package com.system.fcfs.event.dto.request;
+
+public class GetWinnerRequestDTO {
+}
