+++ src/main/java/com/system/fcfs/event/implementation/EventController.java
@@ -1,7 +1,7 @@
-package com.system.fcfs.event.controller;
+package com.system.fcfs.event.implementation;
 
-import com.system.fcfs.event.dto.GetWinnerResponseDTO;
-import com.system.fcfs.event.dto.PostEventRequestDTO;
+import com.system.fcfs.event.dto.response.GetWinnerResponseDTO;
+import com.system.fcfs.event.dto.request.PostEventRequestDTO;
 import com.system.fcfs.event.service.EventService;
 import com.system.fcfs.global.domain.response.CommonResponseEntity;
 import lombok.RequiredArgsConstructor;
@@ -21,17 +21,18 @@ public class EventController {
 
     @PostMapping("/attempt")
     public CommonResponseEntity requestEvent(@RequestBody PostEventRequestDTO postEventRequestDTO) {
-        log.info("Received request: {}", postEventRequestDTO);
+        // log.info("Received request: {}", postEventRequestDTO);
         // 1. 먼저 해당 요청이 유효한지 아닌지 확인한다.
-        couponService.validRequest(postEventRequestDTO);
+        // couponService.validRequest(postEventRequestDTO);
         // 2. 해당 요청을 큐에 넣는다.
         couponService.addQueue(postEventRequestDTO);
         return success(true);
     }
 
-    @GetMapping("/winner")
-    public CommonResponseEntity<List<GetWinnerResponseDTO>> getWinner() {
-        return success(couponService.getTop100Winners());
+    @GetMapping("/winner/{eventName}")
+    public CommonResponseEntity<List<GetWinnerResponseDTO>> getWinner(@PathVariable String eventName) {
+        // log.info("Received request: {}", eventName);
+        return success(couponService.getTop100Winners(eventName));
     }
 }
